refactored newtripform and restyle modal head
