+++ client/src/components/MapElement/MapElement.js
@@ -1,12 +1,8 @@
-import { TextField } from '@mui/material'
 import PropTypes from 'prop-types'
 import { useMemo } from 'react'
-import { useForm } from 'react-hook-form'
-import { useDispatch, useSelector } from 'react-redux'
+import { useSelector } from 'react-redux'
 import { AppView } from '../../constants/enums'
-import { closeNewTripModal } from '../../redux/reducers/modalsSlice'
-import { setAppView } from '../../redux/reducers/viewSlice'
-import { Button, Modal } from '../common'
+import { NewTripForm } from '../TripElement'
 
 MapElement.propTypes = {
   className: PropTypes.string,
@@ -15,79 +11,10 @@ MapElement.propTypes = {
 export function MapElement({ className }) {
   const appView = useSelector((state) => state.view.appView)
   const activeTripId = useSelector((state) => state.view.activeTripId)
-  const newTripModalIsOpen = useSelector(
-    (state) => state.modals.newTripModalIsOpen,
-  )
-  const dispatch = useDispatch()
-
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-    reset,
-  } = useForm()
-
-  const handleCloseNewTripModal = () => {
-    dispatch(closeNewTripModal())
-    reset()
-  }
-
-  const onSubmit = (data) => {
-    // TODO: make api call to create new trip
-    console.log(data)
-    dispatch(setAppView(AppView.TRIP_VIEW))
-    handleCloseNewTripModal()
-  }
-
-  const newTripForm = useMemo(() => {
-    return (
-      <Modal
-        open={appView === AppView.NEW_TRIP && newTripModalIsOpen}
-        handleClose={handleCloseNewTripModal}
-        title='Manifesting A New Trip...'
-        footer={
-          <Button
-            className='mt-4 w-full rounded bg-slate-300 hover:bg-slate-400'
-            onClick={handleSubmit(onSubmit)}
-            type='submit'
-          >
-            Brew 🍵
-          </Button>
-        }
-      >
-        <TextField
-          {...register('destination', { required: true })}
-          label='Destination'
-          placeholder='Tell me where you want to go...'
-          error={!!errors.destination}
-        />
-        <TextField
-          {...register('stagesPerDay', { required: true })}
-          label='Stages per day'
-          placeholder='Tell me how many places you want to visit..'
-          type='number'
-          error={!!errors.stagesPerDay}
-        />
-        <TextField
-          {...register('budget', { required: true })}
-          label='Budget'
-          placeholder='Tell me how much you want to spend...'
-          error={!!errors.budget}
-        />
-        <TextField
-          {...register('numberOfDays', { required: true })}
-          label='Number of days'
-          placeholder='Tell me how long you are going for...'
-          type='number'
-          error={!!errors.numberOfDays}
-        />
-      </Modal>
-    )
-  }, [appView, register, handleSubmit, errors, newTripModalIsOpen])
 
   const mapContent = useMemo(() => {
     if (appView === AppView.NEW_TRIP) {
-      return newTripForm
+      return <NewTripForm />
     }
     return (
       <span className='mt-4 inline-flex w-full flex-row justify-center'>
@@ -96,7 +23,7 @@ export function MapElement({ className }) {
         </h1>
       </span>
     )
-  }, [appView, activeTripId, newTripForm])
+  }, [appView, activeTripId, NewTripForm])
 
   const bgUrl = useMemo(() => {
     if (appView === AppView.NEW_TRIP) return "bg-[url('../public/globe.png')]"
