+++ client/src/components/TripElement/index.js
@@ -1,0 +1,1 @@
+export { NewTripForm } from './NewTripForm'
